uiTestCase/requirements.txt
selenium==4.3.0
pytest==7.1.2
pytest-rerunfailures==10.2
allure-pytest==2.9.45
pyyaml==6.0
openpyxl==3.0.10